Update package now archivew

+++ db/db.go
@@ -28,7 +28,7 @@ func GetPackage(sortingNumber string) (Package, error) {
 
 	// Prepare a statement which gets a package
 	stmt, err := db.Prepare(`
-		SELECT sorting_number, name, building, room, package_type
+		SELECT sorting_number, date_received, name, building, room, package_type
 		FROM Packages
  		WHERE sorting_number = ?`)
 	if err != nil {
@@ -50,7 +50,7 @@ func GetPackage(sortingNumber string) (Package, error) {
 
 		// Store the found row into a variable p
 		var p Package
-		res.Scan(&p.Number, &p.Name, &p.Building, &p.Room, &p.PackageType)
+		res.Scan(&p.Number, &p.DateReceived &p.Name, &p.Building, &p.Room, &p.PackageType)
 		fmt.Printf("%s: %#v\n", sortingNumber, p)
 
 		return p, nil
@@ -88,6 +88,24 @@ func AddPackage(name string, building string, room string, packageType string) (
 func Archive(sortingNumber string) error {
 
 	//Get package information for the archive
+	pickedUpPackage = GetPackage(sortingNumber)
+	stmt, err := db.Prepare("INSERT INTO Picked_Up VALUES(?, ?, ?, ?, ?, ?, DATETIME('now','localtime'), NULL)")
+	if err != nil {
+		log.Println("Error occured while preparing statement:", err)
+		return err
+	}
+	defer stmt.Close()
+	num, err : = stmt.Exec(pickedUpPackage.Number.String(), pickedUpPackage.DateReceived, pickedUpPackage.Name, pickedUpPackage.Building, pickedUpPackage.Room, pickedUpPackage.PackageType)
+	
+	if err != nil {
+		log.Println("Error occured while executing statement:", err)
+		return err
+	}
+
+	rows, _ := num.RowsAffected()
+	if rows == 0 {
+		return ErrNoPackageFound
+	}
 
 	// TODO add a package archive where we store all packages
 	stmt, err := db.Prepare("DELETE FROM Packages WHERE sorting_number=?")
